Move joinable_tables API types
